refactor(logging): move framework logging to separate module
- Moved framework logging functions from `logfiles.py` to new
  `framework_log.py` module.
- Updated imports and references to use the new `framework_log` module.
- Adjusted `temptools.py` to use `IS_XDIST` directly from `os.environ`.
