+++ cardano_node_tests/utils/framework_log.py
@@ -1,0 +1,33 @@
+import logging
+import pathlib as pl
+import time
+
+from cardano_node_tests.utils import helpers
+from cardano_node_tests.utils import temptools
+
+
+@helpers.callonce
+def get_framework_log_path() -> pl.Path:
+    return temptools.get_pytest_worker_tmp() / "framework.log"
+
+
+@helpers.callonce
+def framework_logger() -> logging.Logger:
+    """Get logger for the `framework.log` file.
+
+    The logger is configured per worker. It can be used for logging (and later reporting) events
+    like a failure to start a cluster instance.
+    """
+
+    class UTCFormatter(logging.Formatter):
+        converter = time.gmtime
+
+    formatter = UTCFormatter("%(asctime)s %(levelname)s %(message)s")
+    handler = logging.FileHandler(get_framework_log_path())
+    handler.setFormatter(formatter)
+
+    logger = logging.getLogger("framework")
+    logger.setLevel(logging.INFO)
+    logger.addHandler(handler)
+
+    return logger
